Исправил ошибку в методе создания Пациента
