+++ rekinetix-backend/controllers/patient/patient.utilities.js
@@ -1,3 +1,13 @@
+const getPatientData = (data) => ({
+  fullname: data.fullname,
+  birthday: data.birthday,
+  gender: data.gender,
+  height: data.height,
+  weight: data.weight,
+  phone: data.phone,
+  address: data.address,
+});
+
 const reducePlansToPatients = (planDocs) => {
   return planDocs.reduce((acc, planDoc) => {
     plan = planDoc.toObject();
@@ -13,6 +23,6 @@ const reducePlansToPatients = (planDocs) => {
   }, []);
 };
 
-const utilities = { reducePlansToPatients };
+const utilities = { getPatientData, reducePlansToPatients };
 
 module.exports = utilities;
