+++ rekinetix-backend/models/Patient.js
@@ -19,19 +19,15 @@ const PatientSchema = new Schema(
     },
     height: {
       type: Number,
-      required: true,
     },
     weight: {
       type: Number,
-      required: true,
     },
     phone: {
       type: String,
-      required: true,
     },
     address: {
       type: String,
-      required: true,
     },
   },
   { timestamps: true }
